fix: database migration

+++ database/migrations/2024_05_23_144235_create_role_permission_table.php
@@ -16,9 +16,7 @@ return new class extends Migration
         Schema::create('role_permission', function (Blueprint $table) {
             $table->id();
             $table->string('guard_name');
-            // Tạo cột user_id
-            $table->unsignedBigInteger('role_id');
-            // Tạo ràng buộc khóa ngoại
+            $table->unsignedInteger('role_id');
             $table->foreign('role_id')->references('id')->on('roles');
             $table->timestamps();
         });
